+++ tests/test_basic.py
@@ -1657,7 +1657,6 @@ def test_no_setup_after_first_request(app, client):
     def index():
         return "Awesome"
 
-    assert not app.got_first_request
     assert client.get("/").data == b"Awesome"
 
     with pytest.raises(AssertionError) as exc_info: